project/api/requirements.txt
python-3.12.3
tensorflow==2.17.0
fastapi
uvicorn
python-multipart
pillow
tensorflow-serving-api==2.5.0
matplotlib
numpy
